requirements.txt
robotframework-seleniumlibrary==4.5.0
robotframework-databaselibrary==1.2.4
robotframework-requests==0.7.0
robotframework-datadriver==0.3.6
robotframework-pabot==1.8.1
PyMySQL==0.10.0
allure-robotframework==2.8.17
